Simple website scraping

This function is used to scrap one element from a website.

scrap(link, node, clean = FALSE, askRobot = FALSE)

Arguments

link
the link of the webpage to scrap

node
the CSS element to consider, the SelectorGadget tool is highly recommended

clean
logical. Should the function clean the extracted vector or not ? Default is FALSE.

askRobot
logical. Should the function ask the robots.txt if we're allowed or not to scrap the web page ? Default is FALSE.

Value

a character vector

Examples

# \donttest{
# Extracting imdb top 250 movie titles

link <- "https://www.imdb.com/chart/top/"
node <- ".titleColumn a"

scrap(link, node)# }

#>   [1] "The Shawshank Redemption"                                            
#>   [2] "The Godfather"                                                       
#>   [3] "The Godfather: Part II"                                              
#>   [4] "The Dark Knight"                                                     
#>   [5] "12 Angry Men"                                                        
#>   [6] "Schindler's List"                                                    
#>   [7] "The Lord of the Rings: The Return of the King"                       
#>   [8] "Pulp Fiction"                                                        
#>   [9] "Il buono, il brutto, il cattivo"                                     
#>  [10] "The Lord of the Rings: The Fellowship of the Ring"                   
#>  [11] "Fight Club"                                                          
#>  [12] "Forrest Gump"                                                        
#>  [13] "Inception"                                                           
#>  [14] "Star Wars: Episode V - The Empire Strikes Back"                      
#>  [15] "The Lord of the Rings: The Two Towers"                               
#>  [16] "The Matrix"                                                          
#>  [17] "Goodfellas"                                                          
#>  [18] "One Flew Over the Cuckoo's Nest"                                     
#>  [19] "Shichinin no samurai"                                                
#>  [20] "Se7en"                                                               
#>  [21] "Cidade de Deus"                                                      
#>  [22] "La vita è bella"                                                     
#>  [23] "The Silence of the Lambs"                                            
#>  [24] "It's a Wonderful Life"                                               
#>  [25] "Star Wars"                                                           
#>  [26] "Gisaengchung"                                                        
#>  [27] "Saving Private Ryan"                                                 
#>  [28] "Sen to Chihiro no kamikakushi"                                       
#>  [29] "The Green Mile"                                                      
#>  [30] "Interstellar"                                                        
#>  [31] "Léon"                                                                
#>  [32] "The Usual Suspects"                                                  
#>  [33] "Seppuku"                                                             
#>  [34] "Joker"                                                               
#>  [35] "The Lion King"                                                       
#>  [36] "American History X"                                                  
#>  [37] "Back to the Future"                                                  
#>  [38] "Terminator 2: Judgment Day"                                          
#>  [39] "The Pianist"                                                         
#>  [40] "Modern Times"                                                        
#>  [41] "Psycho"                                                              
#>  [42] "Gladiator"                                                           
#>  [43] "City Lights"                                                         
#>  [44] "The Intouchables"                                                    
#>  [45] "The Departed"                                                        
#>  [46] "Whiplash"                                                            
#>  [47] "Once Upon a Time in the West"                                        
#>  [48] "The Prestige"                                                        
#>  [49] "1917"                                                                
#>  [50] "Casablanca"                                                          
#>  [51] "Hotaru no haka"                                                      
#>  [52] "Rear Window"                                                         
#>  [53] "Nuovo Cinema Paradiso"                                               
#>  [54] "Alien"                                                               
#>  [55] "Raiders of the Lost Ark"                                             
#>  [56] "Apocalypse Now"                                                      
#>  [57] "Memento"                                                             
#>  [58] "The Great Dictator"                                                  
#>  [59] "The Lives of Others"                                                 
#>  [60] "Django Unchained"                                                    
#>  [61] "Avengers: Infinity War"                                              
#>  [62] "Avengers: Endgame"                                                   
#>  [63] "The Shining"                                                         
#>  [64] "Paths of Glory"                                                      
#>  [65] "WALL·E"                                                              
#>  [66] "Spider-Man: Into the Spider-Verse"                                   
#>  [67] "Sunset Blvd."                                                        
#>  [68] "Mononoke-hime"                                                       
#>  [69] "Dr. Strangelove or: How I Learned to Stop Worrying and Love the Bomb"
#>  [70] "Oldeuboi"                                                            
#>  [71] "Witness for the Prosecution"                                         
#>  [72] "The Dark Knight Rises"                                               
#>  [73] "Once Upon a Time in America"                                         
#>  [74] "Aliens"                                                              
#>  [75] "Kimi no na wa."                                                      
#>  [76] "American Beauty"                                                     
#>  [77] "Coco"                                                                
#>  [78] "Braveheart"                                                          
#>  [79] "Das Boot"                                                            
#>  [80] "Tengoku to jigoku"                                                   
#>  [81] "3 Idiots"                                                            
#>  [82] "Star Wars: Episode VI - Return of the Jedi"                          
#>  [83] "Toy Story"                                                           
#>  [84] "Taare Zameen Par"                                                    
#>  [85] "Amadeus"                                                             
#>  [86] "Reservoir Dogs"                                                      
#>  [87] "Inglourious Basterds"                                                
#>  [88] "Good Will Hunting"                                                   
#>  [89] "2001: A Space Odyssey"                                               
#>  [90] "M - Eine Stadt sucht einen Mörder"                                   
#>  [91] "Requiem for a Dream"                                                 
#>  [92] "Vertigo"                                                             
#>  [93] "Dangal"                                                              
#>  [94] "Eternal Sunshine of the Spotless Mind"                               
#>  [95] "Citizen Kane"                                                        
#>  [96] "Jagten"                                                              
#>  [97] "Full Metal Jacket"                                                   
#>  [98] "North by Northwest"                                                  
#>  [99] "A Clockwork Orange"                                                  
#> [100] "Snatch"                                                              
#> [101] "The Kid"                                                             
#> [102] "Ladri di biciclette"                                                 
#> [103] "Scarface"                                                            
#> [104] "Singin' in the Rain"                                                 
#> [105] "Amélie"                                                              
#> [106] "Taxi Driver"                                                         
#> [107] "Lawrence of Arabia"                                                  
#> [108] "Toy Story 3"                                                         
#> [109] "The Sting"                                                           
#> [110] "Capharnaüm"                                                          
#> [111] "Metropolis"                                                          
#> [112] "Per qualche dollaro in più"                                          
#> [113] "Jodaeiye Nader az Simin"                                             
#> [114] "Ikiru"                                                               
#> [115] "Double Indemnity"                                                    
#> [116] "To Kill a Mockingbird"                                               
#> [117] "The Apartment"                                                       
#> [118] "Indiana Jones and the Last Crusade"                                  
#> [119] "Incendies"                                                           
#> [120] "Up"                                                                  
#> [121] "L.A. Confidential"                                                   
#> [122] "Monty Python and the Holy Grail"                                     
#> [123] "Heat"                                                                
#> [124] "Die Hard"                                                            
#> [125] "Rashômon"                                                            
#> [126] "Yôjinbô"                                                             
#> [127] "Batman Begins"                                                       
#> [128] "Unforgiven"                                                          
#> [129] "Green Book"                                                          
#> [130] "Der Untergang"                                                       
#> [131] "Bacheha-Ye aseman"                                                   
#> [132] "Some Like It Hot"                                                    
#> [133] "Hauru no ugoku shiro"                                                
#> [134] "Tonari no Totoro"                                                    
#> [135] "Idi i smotri"                                                        
#> [136] "The Great Escape"                                                    
#> [137] "Ran"                                                                 
#> [138] "All About Eve"                                                       
#> [139] "A Beautiful Mind"                                                    
#> [140] "Casino"                                                              
#> [141] "Pan's Labyrinth"                                                     
#> [142] "Raging Bull"                                                         
#> [143] "El secreto de sus ojos"                                              
#> [144] "Lock, Stock and Two Smoking Barrels"                                 
#> [145] "The Wolf of Wall Street"                                             
#> [146] "The Treasure of the Sierra Madre"                                    
#> [147] "Judgment at Nuremberg"                                               
#> [148] "Three Billboards Outside Ebbing, Missouri"                           
#> [149] "Babam ve Oglum"                                                      
#> [150] "Chinatown"                                                           
#> [151] "There Will Be Blood"                                                 
#> [152] "The Gold Rush"                                                       
#> [153] "Ford v Ferrari"                                                      
#> [154] "Dial M for Murder"                                                   
#> [155] "V for Vendetta"                                                      
#> [156] "Inside Out"                                                          
#> [157] "Det sjunde inseglet"                                                 
#> [158] "Warrior"                                                             
#> [159] "No Country for Old Men"                                              
#> [160] "Trainspotting"                                                       
#> [161] "The Elephant Man"                                                    
#> [162] "Shutter Island"                                                      
#> [163] "Room"                                                                
#> [164] "The Sixth Sense"                                                     
#> [165] "The Thing"                                                           
#> [166] "Gone with the Wind"                                                  
#> [167] "Blade Runner"                                                        
#> [168] "The Bridge on the River Kwai"                                        
#> [169] "Jurassic Park"                                                       
#> [170] "The Third Man"                                                       
#> [171] "On the Waterfront"                                                   
#> [172] "Finding Nemo"                                                        
#> [173] "Smultronstället"                                                     
#> [174] "Fargo"                                                               
#> [175] "Andhadhun"                                                           
#> [176] "Gran Torino"                                                         
#> [177] "Kill Bill: Vol. 1"                                                   
#> [178] "The Deer Hunter"                                                     
#> [179] "Eskiya"                                                              
#> [180] "Stalker"                                                             
#> [181] "Tôkyô monogatari"                                                    
#> [182] "The Truman Show"                                                     
#> [183] "Relatos salvajes"                                                    
#> [184] "Klaus"                                                               
#> [185] "The Big Lebowski"                                                    
#> [186] "Mary and Max"                                                        
#> [187] "Salinui chueok"                                                      
#> [188] "In the Name of the Father"                                           
#> [189] "Gone Girl"                                                           
#> [190] "Hacksaw Ridge"                                                       
#> [191] "Mr. Smith Goes to Washington"                                        
#> [192] "The General"                                                         
#> [193] "How to Train Your Dragon"                                            
#> [194] "The Grand Budapest Hotel"                                            
#> [195] "Sherlock Jr."                                                        
#> [196] "Before Sunrise"                                                      
#> [197] "Persona"                                                             
#> [198] "Catch Me If You Can"                                                 
#> [199] "Prisoners"                                                           
#> [200] "12 Years a Slave"                                                    
#> [201] "Cool Hand Luke"                                                      
#> [202] "Into the Wild"                                                       
#> [203] "Network"                                                             
#> [204] "Le salaire de la peur"                                               
#> [205] "Stand by Me"                                                         
#> [206] "Life of Brian"                                                       
#> [207] "Mad Max: Fury Road"                                                  
#> [208] "Barry Lyndon"                                                        
#> [209] "Platoon"                                                             
#> [210] "Million Dollar Baby"                                                 
#> [211] "Hachi: A Dog's Tale"                                                 
#> [212] "La passion de Jeanne d'Arc"                                          
#> [213] "Rush"                                                                
#> [214] "Logan"                                                               
#> [215] "Ben-Hur"                                                             
#> [216] "Rang De Basanti"                                                     
#> [217] "Andrei Rublev"                                                       
#> [218] "Kaze no tani no Naushika"                                            
#> [219] "Harry Potter and the Deathly Hallows: Part 2"                        
#> [220] "Les quatre cents coups"                                              
#> [221] "Dead Poets Society"                                                  
#> [222] "Hotel Rwanda"                                                        
#> [223] "Amores perros"                                                       
#> [224] "Spotlight"                                                           
#> [225] "Rebecca"                                                             
#> [226] "The Red Shoes"                                                       
#> [227] "Rocky"                                                               
#> [228] "Marriage Story"                                                      
#> [229] "Ah-ga-ssi"                                                           
#> [230] "Monsters, Inc."                                                      
#> [231] "Ace in the Hole"                                                     
#> [232] "La haine"                                                            
#> [233] "It Happened One Night"                                               
#> [234] "White Heat"                                                          
#> [235] "The Princess Bride"                                                  
#> [236] "Lagaan: Once Upon a Time in India"                                   
#> [237] "Gangs of Wasseypur"                                                  
#> [238] "Before Sunset"                                                       
#> [239] "Faa yeung nin wa"                                                    
#> [240] "The Help"                                                            
#> [241] "The Terminator"                                                      
#> [242] "Butch Cassidy and the Sundance Kid"                                  
#> [243] "Paris, Texas"                                                        
#> [244] "PK"                                                                  
#> [245] "Akira"                                                               
#> [246] "Contratiempo"                                                        
#> [247] "Aladdin"                                                             
#> [248] "Tenkû no shiro Rapyuta"                                              
#> [249] "Swades: We, the People"                                              
#> [250] "Shin seiki Evangelion Gekijô-ban: Air/Magokoro wo, kimi ni"