Website title scraping

This function is used to scrape titles (h1, h2 & h3 html tags) from a website. Useful for scraping daily electronic newspapers' titles.

titles_scrap(link, contain = NULL, case_sensitive = FALSE, askRobot = FALSE)

Arguments

link
the link of the web page to scrape

contain
filter the titles according to a character string provided.

case_sensitive
logical. Should the contain argument be case sensitive ? defaults to FALSE

askRobot
logical. Should the function ask the robots.txt if we're allowed or not to scrape the web page ? Default is FALSE

Value

a character vector

Examples

# \donttest{
# Extracting the current titles of the New York Times

link     <- "https://www.nytimes.com/"

titles_scrap(link)# }

#>  [1] " Live Updates "                                                                                                                                                                                                                                                                                                           
#>  [2] " See full presidential results "                                                                                                                                                                                                                                                                                          
#>  [3] " Key States "                                                                                                                                                                                                                                                                                                             
#>  [4] "U.S. Surpasses 10 Million Coronavirus Cases"                                                                                                                                                                                                                                                                              
#>  [5] "An Explanation for Some Covid-19 Deaths May Not Be Holding Up"                                                                                                                                                                                                                                                            
#>  [6] "Alex Trebek, Longtime Host of ‘Jeopardy!,’ Dies at 80"                                                                                                                                                                                                                                                                    
#>  [7] "O Captain, My ‘Jeopardy!’ Captain"                                                                                                                                                                                                                                                                                        
#>  [8] "“We have no idea what the show is going to be without him.” Fans mourned the loss of Alex Trebek."                                                                                                                                                                                                                        
#>  [9] "Making It Work"                                                                                                                                                                                                                                                                                                           
#> [10] "Yes, it is possible to buy a good used computer online. Here’s how."                                                                                                                                                                                                                                                      
#> [11] "If you’re on edge, try taking a controlled breath."                                                                                                                                                                                                                                                                       
#> [12] "The Weekender"                                                                                                                                                                                                                                                                                                            
#> [13] "Did you follow the news this week? Take our quiz."                                                                                                                                                                                                                                                                        
#> [14] "We Hereby Dump Trump"                                                                                                                                                                                                                                                                                                     
#> [15] "What Alex Trebek Was Really Like"                                                                                                                                                                                                                                                                                         
#> [16] "Black Empowerment Outside the Headlines"                                                                                                                                                                                                                                                                                  
#> [17] "The Human Experience Will Not Be Quantified"                                                                                                                                                                                                                                                                              
#> [18] "Elections Don’t Have to Be So Chaotic and Excruciating"                                                                                                                                                                                                                                                                   
#> [19] "A Loving Chastisement for America"                                                                                                                                                                                                                                                                                        
#> [20] "Victory for Joe Biden, at Last"                                                                                                                                                                                                                                                                                           
#> [21] "Is There a Trumpism After Trump?"                                                                                                                                                                                                                                                                                         
#> [22] "A Black Hero in the Jim Crow Navy"                                                                                                                                                                                                                                                                                        
#> [23] "Biden Can’t Be F.D.R. He Could Still Be L.B.J."                                                                                                                                                                                                                                                                           
#> [24] "Why You Should Brave the ‘Bad’ Weather"                                                                                                                                                                                                                                                                                   
#> [25] "How Do You Know When Society Is About to Fall Apart?"                                                                                                                                                                                                                                                                     
#> [26] "‘Saturday Night Live’ Sends Off Trump With a Ballad"                                                                                                                                                                                                                                                                      
#> [27] "Review: At Home With Our Ancient Cousins, the Neanderthals"                                                                                                                                                                                                                                                               
#> [28] "Site Index"                                                                                                                                                                                                                                                                                                               
#> [29] "Site Information Navigation"                                                                                                                                                                                                                                                                                              
#> [30] "As Biden Plans for Day 1, G.O.P. Navigates Trump’s Refusal to Concede"                                                                                                                                                                                                                                                    
#> [31] "George W. Bush Congratulates Biden"                                                                                                                                                                                                                                                                                       
#> [32] ".css-11f6h48{margin-right:5px;}.css-11f6h48:hover{border-bottom:1px solid #e2e2e2;}Putin is silent on Biden’s victory, foreshadowing tense years ahead..css-m7dmkp{font-family:nyt-franklin,helvetica,arial,sans-serif;color:#d0021b;font-weight:600;font-size:0.75rem;line-height:0.8125rem;display:inline-block;}1h ago"
#> [33] "Christian conservatives reflect on Trump’s loss and look ahead.1h ago"                                                                                                                                                                                                                                                    
#> [34] "Here are 5 things the Biden administration could do quickly on the environment.2h ago"                                                                                                                                                                                                                                    
#> [35] "Biden’s German shepherds will restore a presidential tradition.2h ago"                                                                                                                                                                                                                                                    
#> [36] "President"                                                                                                                                                                                                                                                                                                                
#> [37] "Senate"                                                                                                                                                                                                                                                                                                                   
#> [38] "House"                                                                                                                                                                                                                                                                                                                    
#> [39] "Senate"                                                                                                                                                                                                                                                                                                                   
#> [40] "House"                                                                                                                                                                                                                                                                                                                    
#> [41] "Five Takeaways From Biden’s Victory Speech"                                                                                                                                                                                                                                                                               
#> [42] "Why the Biden Economy Could Be the Same Long Slog as the Obama Economy"                                                                                                                                                                                                                                                   
#> [43] "‘A Long Time Coming’: Black Women Celebrate Harris’s Ascension"                                                                                                                                                                                                                                                           
#> [44] "Kamala Harris’s ancestral village in India rejoiced, but many wondered what changes may come."                                                                                                                                                                                                                            
#> [45] "What’s Next for Trump? Family Business Awaits His Return"                                                                                                                                                                                                                                                                 
#> [46] "Analysis: Dancing in the Streets for Joe Biden"                                                                                                                                                                                                                                                                           
#> [47] "A Crisis Candidacy, Then Party Unity"                                                                                                                                                                                                                                                                                     
#> [48] "As a Bitter Election Ends, 26 Americans Look to the Future"                                                                                                                                                                                                                                                               
#> [49] "For 9 of 10 Voters, Protests Over Police Violence Played Role in Their Choice"                                                                                                                                                                                                                                            
#> [50] "Georgia, From Reliably Republican to Political Ground Zero"                                                                                                                                                                                                                                                               
#> [51] "‘It’s Such a Relief’: Biden Voters Rebuild a Wall That Trump Smashed"                                                                                                                                                                                                                                                     
#> [52] "How Biden Flipped Pennsylvania and Won the Election"                                                                                                                                                                                                                                                                      
#> [53] "With control of the Senate unclear, liberal Democrats are lining up to press their agenda."                                                                                                                                                                                                                               
#> [54] "AT HOME"                                                                                                                                                                                                                                                                                                                  
#> [55] "Opinion"                                                                                                                                                                                                                                                                                                                  
#> [56] "Editors’ Picks"                                                                                                                                                                                                                                                                                                           
#> [57] "Advertisement"