Website text paragraph scraping

This function is used to scrape text paragraphs from a website.

paragraphs_scrap(
  link,
  contain = NULL,
  case_sensitive = FALSE,
  collapse = FALSE,
  askRobot = FALSE
)

Arguments

link
the link of the web page to scrape

contain
filter the paragraphs according to the character string provided.

case_sensitive
logical. Should the contain argument be case sensitive ? defaults to FALSE

collapse
if TRUE the paragraphs will be collapsed into one element and the contain argument ignored.

askRobot
logical. Should the function ask the robots.txt if we're allowed or not to scrap the web page ? Default is FALSE.

Value

a character vector.

Examples

# \donttest{
# Extracting the paragraphs displayed on the health page of the New York Times

link     <- "https://www.nytimes.com/section/health"

paragraphs_scrap(link)# }

#>  [1] "Advertisement"                                                                                                                                                                                                                        
#>  [2] "Supported by"                                                                                                                                                                                                                         
#>  [3] "Appointed head of the incoming administration’s task force on health equity, the Yale University scientist “is not sitting in her ivory tower.”"                                                                                      
#>  [4] "By Roni Caryn Rabin"                                                                                                                                                                                                                  
#>  [5] "A large study of patients from a Wuhan, China hospital showed that a half-year later, three-quarters were struggling with problems like fatigue, depression and diminished lung function."                                            
#>  [6] "By Pam Belluck"                                                                                                                                                                                                                       
#>  [7] "Reports of a highly contagious new variant, published on Friday by multiple news outlets, were based on speculative statements made by Dr. Deborah Birx."                                                                             
#>  [8] "By Apoorva Mandavilli"                                                                                                                                                                                                                
#>  [9] "It’s good news, but experts cautioned that the new variants from Britain and South Africa also carry other potentially dangerous mutations that have not yet been investigated."                                                      
#> [10] "By Carl Zimmer"                                                                                                                                                                                                                       
#> [11] "The big dip in mortality shown in the trial of about 800 patients has caught some experts by surprise because previous studies of the drugs showed little benefit."                                                                   
#> [12] "By Katherine J. Wu"                                                                                                                                                                                                                   
#> [13] "An outbreak aboard a September flight from Dubai to New Zealand offers researchers, and airlines, an opportunity to study in-transit contagion."                                                                                      
#> [14] "By Benedict Carey"                                                                                                                                                                                                                    
#> [15] "Some scientists fear that the mayhem on Capitol Hill may prove to have been a so-called super-spreading event."                                                                                                                       
#> [16] "By Apoorva Mandavilli"                                                                                                                                                                                                                
#> [17] "It’s not too late to curb the contagious variant’s spread in the U.S., experts say — but only with a national program for genetic sequencing."                                                                                        
#> [18] "By Carl Zimmer"                                                                                                                                                                                                                       
#> [19] "See how the leading vaccines train the immune system to fight the virus."                                                                                                                                                             
#> [20] "By Jonathan Corum and Carl Zimmer"                                                                                                                                                                                                    
#> [21] "Advertisement"                                                                                                                                                                                                                        
#> [22] "Whether it’s a person or an app that sends us reminders, we make better choices when we’re being watched (even by ourselves.)"                                                                                                        
#> [23] "By Tara Parker-Pope"                                                                                                                                                                                                                  
#> [24] "An important lesson of 2020 is that preparation for a health crisis is a gift to the people you love. Today’s Well challenge shows you how."                                                                                          
#> [25] "By Tara Parker-Pope"                                                                                                                                                                                                                  
#> [26] "Our 7-Day Well Challenge will show you how to build on the healthy habits you learned during pandemic life."                                                                                                                          
#> [27] "By Tara Parker-Pope"                                                                                                                                                                                                                  
#> [28] "During pandemic lockdowns, many of us learned the importance of short home workouts. Take the 7-Day Well Challenge for a new exercise video and more ways to keep moving in 2021."                                                    
#> [29] "By Tara Parker-Pope"                                                                                                                                                                                                                  
#> [30] "Social bubbles have helped us cope with the restrictions of Covid-19. For today’s Well Challenge, look to your pandemic pod to inspire and motivate you toward a healthier and happier life."                                         
#> [31] "By Tara Parker-Pope"                                                                                                                                                                                                                  
#> [32] "Critical eviction moratoriums don’t offer much relief for mom-and-pop landlords who still have to pay to keep their buildings operating."                                                                                             
#> [33] "By Ronda Kaysen"                                                                                                                                                                                                                      
#> [34] "Blocked from obtaining vaccines from the United States, its putative ally, and taunted in Russian propaganda, Ukraine turned to China."                                                                                               
#> [35] "By Maria Varenikova"                                                                                                                                                                                                                  
#> [36] "No, not Tom Brady. He made it. But the pandemic is forcing teams to keep large numbers of fans away."                                                                                                                                 
#> [37] "By Ken Belson"                                                                                                                                                                                                                        
#> [38] "El fuerte declive en la afluencia de visitantes desde el inicio de la pandemia presionó a esta pequeña comunidad, ubicada en las colinas de Chianti, a aferrarse a lo esencial: la farmacia, la tienda de alimentos y la agricultura."
#> [39] "By Gaia Pianigiani"                                                                                                                                                                                                                   
#> [40] "El gobierno espera emprender pronto la inmunización a nivel nacional, pero la vacuna de la empresa china requiere la aprobación de los reguladores de seguridad y de un influyente consejo religioso."                                
#> [41] "By Richard C. Paddock"                                                                                                                                                                                                                
#> [42] "Advertisement"                                                                                                                                                                                                                        
#> [43] "Removing obstacles makes it more likely you’ll achieve a new health goal. The 7-Day Well Challenge will show you how."                                                                                                                
#> [44] "By Tara Parker-Pope"                                                                                                                                                                                                                  
#> [45] "The Trump administration’s move, which Biden could eventually reverse, would loosen program rules and cap the state’s funding as part of a block grant."                                                                              
#> [46] "By Margot Sanger-Katz"                                                                                                                                                                                                                
#> [47] "Public health officials are using Eventbrite’s online ticketing platform to schedule shots. But appointments are still hard to find, and not everyone knows how to use the website."                                                  
#> [48] "By Jacey Fortin"                                                                                                                                                                                                                      
#> [49] "Many people want a pandemic baby, but some sperm banks are running low. So women are joining unregulated Facebook groups to find willing donors, no middleman required."                                                              
#> [50] "By Nellie Bowles"                                                                                                                                                                                                                     
#> [51] "The babysitter says she has nine days sober, but we all lie, every addict, every alcoholic."                                                                                                                                          
#> [52] "By Sarah Twombly"                                                                                                                                                                                                                     
#> [53] "Test your knowledge of this week’s health news."                                                                                                                                                                                      
#> [54] "By Toby Bilanow"                                                                                                                                                                                                                      
#> [55] "By Roni Caryn Rabin"                                                                                                                                                                                                                  
#> [56] "By Sheri Fink"                                                                                                                                                                                                                        
#> [57] "Two adenoviruses help prime the immune system to fight the coronavirus."                                                                                                                                                              
#> [58] "By Jonathan Corum and Carl Zimmer"                                                                                                                                                                                                    
#> [59] "At the start of the pandemic, I was “Dr. No” to my in-laws and cancer patients, but my conversations have become more nuanced."                                                                                                       
#> [60] "By Mikkael A. Sekeres, M.D."                                                                                                                                                                                                          
#> [61] "Advertisement"                                                                                                                                                                                                                        
#> [62] "Advertisement"