Website text paragraph scraping

This function is used to scrape text paragraphs from a website.

paragraphs_scrap(
  link,
  contain = NULL,
  case_sensitive = FALSE,
  collapse = FALSE,
  askRobot = FALSE
)

Arguments

link
the link of the web page to scrape

contain
filter the paragraphs according to the character string provided.

case_sensitive
logical. Should the contain argument be case sensitive ? defaults to FALSE

collapse
if TRUE the paragraphs will be collapsed into one element and the contain argument ignored.

askRobot
logical. Should the function ask the robots.txt if we're allowed or not to scrap the web page ? Default is FALSE.

Value

a character vector.

Examples

# \donttest{
# Extracting the paragraphs displayed on the health page of the New York Times

link     <- "https://www.nytimes.com/section/health"

paragraphs_scrap(link)# }

#>  [1] "Advertisement"                                                                                                                                                                                                         
#>  [2] "Supported by"                                                                                                                                                                                                          
#>  [3] "Recent studies have created doubts about an agent in cytokine storms, and suggest that treatments for it may not help."                                                                                                
#>  [4] "By Gina Kolata"                                                                                                                                                                                                        
#>  [5] "The drug, aducanumab, made by Biogen, would be the first new Alzheimer’s treatment in nearly two decades. But the advisory panel said there was not enough evidence of its effectiveness in slowing cognitive decline."
#>  [6] "By Pam Belluck"                                                                                                                                                                                                        
#>  [7] "Mark Meadows aided President Trump’s efforts to play down the virus throughout the summer. A Trump campaign adviser and four other White House aides also tested positive."                                            
#>  [8] "By Maggie Haberman and Michael D. Shear"                                                                                                                                                                               
#>  [9] "Scientists at Columbia University have developed a treatment that blocks the virus in the nose and lungs, is inexpensive and needs no refrigeration."                                                                  
#> [10] "By Donald G. McNeil Jr."                                                                                                                                                                                               
#> [11] "The research suggests that children clear the infection much faster than adults and may help explain why many don’t become seriously ill."                                                                             
#> [12] "By Apoorva Mandavilli"                                                                                                                                                                                                 
#> [13] "Three distributors and a drug manufacturer have proposed a deal that a majority of states and negotiators for small governments finally seem to like."                                                                 
#> [14] "By Jan Hoffman"                                                                                                                                                                                                        
#> [15] "Government officials said on Wednesday that a mutation in the virus could interfere with vaccine effectiveness in humans."                                                                                             
#> [16] "By James Gorman"                                                                                                                                                                                                       
#> [17] "If symptomatic, they were more likely to develop complications and die than nonpregnant women with symptoms."                                                                                                          
#> [18] "By Roni Caryn Rabin"                                                                                                                                                                                                   
#> [19] "A head-to-head comparison of lab and rapid coronavirus tests drew mixed reactions from experts, who raised concerns about accuracy."                                                                                   
#> [20] "By Katherine J. Wu"                                                                                                                                                                                                    
#> [21] "Advertisement"                                                                                                                                                                                                         
#> [22] "Public health officials urge families to keep celebrations small, avoid mixing households and open the windows."                                                                                                       
#> [23] "By Tara Parker-Pope"                                                                                                                                                                                                   
#> [24] "A basic program of lunges, lifts, squats and crunches eased anxiety levels in healthy young men and women."                                                                                                            
#> [25] "By Gretchen Reynolds"                                                                                                                                                                                                  
#> [26] "Athletes who have endured the most grueling tests have a lot to tell us about how to thrive in the pandemic."                                                                                                          
#> [27] "By Talya Minsberg"                                                                                                                                                                                                     
#> [28] "If the one-two punch of pandemic stress and election stress feels like more than you can handle, try these tips to help you cope."                                                                                     
#> [29] "By Tara Parker-Pope"                                                                                                                                                                                                   
#> [30] "This winter the pandemic is expected to intensify the depression experienced by many people with the syndrome known as seasonal affective disorder, or SAD."                                                           
#> [31] "By Jane E. Brody"                                                                                                                                                                                                      
#> [32] "It’s President-elect Joe Biden."                                                                                                                                                                                       
#> [33] "By Carole Landry"                                                                                                                                                                                                      
#> [34] "Mink in Denmark are not the only animals that could become reservoirs for the coronavirus to spread new mutations to people."                                                                                          
#> [35] "By James Gorman"                                                                                                                                                                                                       
#> [36] "Here’s an overview on President-elect Biden’s positions on coronavirus, health care, the economy, taxes and climate change."                                                                                           
#> [37] "By Sydney Ember"                                                                                                                                                                                                       
#> [38] "We asked voters about their hopes for the next four years — regardless of the occupant in the White House."                                                                                                            
#> [39] "By The New York Times"                                                                                                                                                                                                 
#> [40] "Los emoticonos y su evolución, los emojis, representan el mundo tal y como lo entendemos en 2020."                                                                                                                     
#> [41] "By Jorge Carrión"                                                                                                                                                                                                      
#> [42] "Advertisement"                                                                                                                                                                                                         
#> [43] "Thanksgiving is even more of a logistical puzzle this year, and some people are solving it differently."                                                                                                               
#> [44] "By Courtney Rubin"                                                                                                                                                                                                     
#> [45] "Single people make up one third of all American households and  finding ways to celebrate this Thanksgiving means taking action ahead of time."                                                                        
#> [46] "By Anna Goldfarb"                                                                                                                                                                                                      
#> [47] "Plan out the trip home and set ground rules for safe socializing. And remember that they’ve had a tough year, too, so give them the space to talk about it."                                                           
#> [48] "By Caren Osten Gerszberg"                                                                                                                                                                                              
#> [49] "With the family together for the holiday, you might be looking for tabletop games that will delight kids and grown-ups alike. Here are a few you can all play together."                                               
#> [50] "By Alexis Soloski"                                                                                                                                                                                                     
#> [51] "If you’re shopping for health insurance this fall for coverage starting Jan. 1, start early and set aside a full day to learn about your options."                                                                     
#> [52] "By Joseph Burns"                                                                                                                                                                                                       
#> [53] "This week, pick the best indoor plants, jam with the ‘queen of Kindie rock,’ dive deeper into the power of the X-ray and be swept away by the San Francisco Symphony."                                                 
#> [54] "By Adriana Balsamo and Katherine Cusumano"                                                                                                                                                                             
#> [55] "For my son, each birthday has been accompanied by complicated feelings, a reminder of the milestones he has not met and may never meet."                                                                               
#> [56] "By Alysia Abbott"                                                                                                                                                                                                      
#> [57] "A surprise charge that can take advantage of vulnerable people and possibly violate consumer protection laws."                                                                                                         
#> [58] "By Sarah Kliff and Jessica Silver-Greenberg"                                                                                                                                                                           
#> [59] "It depends on a complicated interplay of transmission rates, available doses and societal goals."                                                                                                                      
#> [60] "By Kim Tingley"                                                                                                                                                                                                        
#> [61] "Scientists are divided on broader implications of the find for ancient gender roles."                                                                                                                                  
#> [62] "By James Gorman"                                                                                                                                                                                                       
#> [63] "Advertisement"                                                                                                                                                                                                         
#> [64] "Advertisement"