Website text paragraph scraping

This function is used to scrape text paragraphs from a website.

paragraphs_scrap(
  link,
  contain = NULL,
  case_sensitive = FALSE,
  collapse = FALSE,
  askRobot = FALSE
)

Arguments

link
the link of the web page to scrape

contain
filter the paragraphs according to the character string provided.

case_sensitive
logical. Should the contain argument be case sensitive ? defaults to FALSE

collapse
if TRUE the paragraphs will be collapsed into one element and the contain argument ignored.

askRobot
logical. Should the function ask the robots.txt if we're allowed or not to scrap the web page ? Default is FALSE.

Value

a character vector.

Examples

# \donttest{
# Extracting the paragraphs displayed on the health page of the New York Times

link     <- "https://www.nytimes.com/section/health"

paragraphs_scrap(link)# }

#>  [1] "Advertisement"                                                                                                                                                                                                                  
#>  [2] "Supported by"                                                                                                                                                                                                                   
#>  [3] "Health officials and hospitals are struggling with a lack of resources. Holiday staffing and saving doses for nursing homes are also contributing to delays."                                                                   
#>  [4] "By Rebecca Robbins, Frances Robles and Tim Arango"                                                                                                                                                                              
#>  [5] "If a second dose of one vaccine isn’t available, another may be substituted, according to new U.K. guidelines."                                                                                                                 
#>  [6] "By Katherine J. Wu"                                                                                                                                                                                                             
#>  [7] "Many Black and Hispanic Americans mistrust government officials, and instead have turned to physicians they have long known."                                                                                                   
#>  [8] "By Gina Kolata"                                                                                                                                                                                                                 
#>  [9] "Reports of Covid-19 cases that appeared shortly after a single shot of a two-dose vaccine shouldn’t cause concern."                                                                                                             
#> [10] "By Katherine J. Wu"                                                                                                                                                                                                             
#> [11] "Contagiousness is the hallmark of the mutated virus surfacing in the U.S. and more than a dozen other countries."                                                                                                               
#> [12] "By Apoorva Mandavilli"                                                                                                                                                                                                          
#> [13] "A more contagious version of the coronavirus may alter the course of the pandemic in the United States, researchers said."                                                                                                      
#> [14] "By Apoorva Mandavilli"                                                                                                                                                                                                          
#> [15] "Significant numbers of coronavirus patients experience long-term symptoms that send them back to the hospital, taxing an already overburdened health system."                                                                   
#> [16] "By Pam Belluck"                                                                                                                                                                                                                 
#> [17] "Knowing the amount of virus carried in the body could help doctors predict the course of a patient’s illness."                                                                                                                  
#> [18] "By Apoorva Mandavilli"                                                                                                                                                                                                          
#> [19] "Most had no history of mental illness and became psychotic weeks after contracting the virus. Cases are expected to remain rare but are being reported worldwide."                                                              
#> [20] "By Pam Belluck"                                                                                                                                                                                                                 
#> [21] "Advertisement"                                                                                                                                                                                                                  
#> [22] "Our 7-Day Well Challenge will show you how to build on the healthy habits you learned during pandemic life."                                                                                                                    
#> [23] "By Tara Parker-Pope"                                                                                                                                                                                                            
#> [24] "2020 was a rough one. Here’s how to make goals for 2021 that feel both satisfying and doable."                                                                                                                                  
#> [25] "By Christina Caron"                                                                                                                                                                                                             
#> [26] "Four seconds of intense intervals, repeated until they amount to a minute of total exertion, led to rapid improvements in strength and fitness in middle-aged and older adults."                                                
#> [27] "By Gretchen Reynolds"                                                                                                                                                                                                           
#> [28] "How will you retain gratitude for the return to normal life? “I’ve come from 2020 to remind you not to forget,” my daughter wrote to her future self."                                                                          
#> [29] "By Julia Cho"                                                                                                                                                                                                                   
#> [30] "Test your knowledge of this week’s health news."                                                                                                                                                                                
#> [31] "By Toby Bilanow"                                                                                                                                                                                                                
#> [32] "New Year’s Eve was mellow for most New Yorkers, and Times Square was practically empty. But some could not resist the lure of a big night out."                                                                                 
#> [33] "By Michael Wilson"                                                                                                                                                                                                              
#> [34] "Badly hit by the coronavirus, Israel has distributed the first of two vaccine doses to more than 10 percent of its population. Prime Minister Netanyahu is leading the charge, bolstering his own battered image along the way."
#> [35] "By Isabel Kershner"                                                                                                                                                                                                             
#> [36] "States and counties are finding it hard to keep jails and prisons open as the virus ravages both prisoners and staff members. But transferring inmates can spur new outbreaks."                                                 
#> [37] "By Brendon Derr, Rebecca Griesbach and Danya Issawi"                                                                                                                                                                            
#> [38] "He was known as one of the great romantic composers. His songs were performed by Elvis Presley, Andrea Bocelli, Christina Aguilera and many others."                                                                            
#> [39] "By Oscar Lopez"                                                                                                                                                                                                                 
#> [40] "As the coronavirus invades locker rooms and postpones matches, soccer needs to face up to a potential worst-case scenario."                                                                                                     
#> [41] "By Rory Smith"                                                                                                                                                                                                                  
#> [42] "Advertisement"                                                                                                                                                                                                                  
#> [43] "By Lauren Wolfe"                                                                                                                                                                                                                
#> [44] "By Kate Taylor"                                                                                                                                                                                                                 
#> [45] "By Kate Taylor"                                                                                                                                                                                                                 
#> [46] "“We know that it should be better, and we’re working hard to make it better,” said Moncef Slaoui, a leader of the federal effort to accelerate vaccine development and distribution."                                           
#> [47] "By Rebecca Robbins"                                                                                                                                                                                                             
#> [48] "Each day, our editors collect the most interesting, striking or delightful facts to appear in articles throughout the paper. Here are 74 from the past year that were the most revealing."                                      
#> [49] "By Times Insider Staff and Clay Hickson"                                                                                                                                                                                        
#> [50] "Using inactivated coronaviruses to train the immune system."                                                                                                                                                                    
#> [51] "By Jonathan Corum and Carl Zimmer"                                                                                                                                                                                              
#> [52] "Using a coronavirus protein to train the immune system."                                                                                                                                                                        
#> [53] "By Jonathan Corum and Carl Zimmer"                                                                                                                                                                                              
#> [54] "White Americans have higher rates than most other racial and ethnic groups."                                                                                                                                                    
#> [55] "By Austin Frakt"                                                                                                                                                                                                                
#> [56] "Health officials hope to soon vaccinate up to two million people per week as the country’s hospitals are overwhelmed by cases of a new, more contagious coronavirus variant."                                                   
#> [57] "By Benjamin Mueller and Rebecca Robbins"                                                                                                                                                                                        
#> [58] "The fact that the infected man had no travel history raises the worrisome possibility that the variant is already well established in his community — and perhaps elsewhere in the United States."                              
#> [59] "By Carl Zimmer and Bryan Pietsch"                                                                                                                                                                                               
#> [60] "Advertisement"                                                                                                                                                                                                                  
#> [61] "Advertisement"