Website text paragraph scraping

This function is used to scrape text paragraphs from a website.

paragraphs_scrap(
  link,
  contain = NULL,
  case_sensitive = FALSE,
  collapse = FALSE,
  askRobot = FALSE
)

Arguments

link

the link of the web page to scrape

contain

filter the paragraphs according to the character string provided.

case_sensitive

logical. Should the contain argument be case sensitive ? defaults to FALSE

collapse

if TRUE the paragraphs will be collapsed into one element and the contain argument ignored.

askRobot

logical. Should the function ask the robots.txt if we're allowed or not to scrap the web page ? Default is FALSE.

Value

a character vector.

Examples

# \donttest{
# Extracting the paragraphs displayed on the health page of the New York Times

link     <- "https://www.nytimes.com/section/health"

paragraphs_scrap(link)# }
#>  [1] "Advertisement"                                                                                                                                                                                                                                                                         
#>  [2] "Supported by"                                                                                                                                                                                                                                                                          
#>  [3] "Significant numbers of older people have the condition. Many find relief with an effective treatment that is being more widely prescribed."                                                                                                                                            
#>  [4] " By Paula Span"                                                                                                                                                                                                                                                                        
#>  [5] "He walked away from his family’s hugely successful ice cream business to crusade for a plant-based diet and against cruelty to animals."                                                                                                                                               
#>  [6] " By Jeré Longman"                                                                                                                                                                                                                                                                      
#>  [7] "Critics saw in the move the beginnings of a more restrictive approach to providing vaccines to Americans."                                                                                                                                                                             
#>  [8] " By Apoorva Mandavilli"                                                                                                                                                                                                                                                                
#>  [9] "The reconstituted C.D.C. panel will revisit the standard vaccination schedule. The former head of an anti-vaccine group is now a special federal employee."                                                                                                                            
#> [10] " By Apoorva Mandavilli"                                                                                                                                                                                                                                                                
#> [11] "Kennedy Withdraws U.S. Funding Pledge to International Vaccine Agency"                                                                                                                                                                                                                 
#> [12] "Robert F. Kennedy Jr. claimed that the agency, Gavi, had “ignored the science” in immunizing children around the world."                                                                                                                                                               
#> [13] " By Stephanie Nolen"                                                                                                                                                                                                                                                                   
#> [14] "Grand Jury Indicts Russian Scientist on Smuggling Charges"                                                                                                                                                                                                                             
#> [15] "Kseniia Petrova, a Harvard researcher, was detained in February after failing to declare scientific samples she was carrying into the country."                                                                                                                                        
#> [16] " By Ellen Barry"                                                                                                                                                                                                                                                                       
#> [17] "N.I.H. Memo Pauses Cancellations of Medical Research Grants"                                                                                                                                                                                                                           
#> [18] "The directive, in a memo issued Tuesday, came after two court rulings that questioned the Trump administration’s swift cuts to funding."                                                                                                                                               
#> [19] " By Benjamin Mueller"                                                                                                                                                                                                                                                                  
#> [20] "The Evolution of Trump’s Views on Foreign Aid"                                                                                                                                                                                                                                         
#> [21] "The administration has gutted agencies like U.S.A.I.D., and President Trump has denigrated their work as wasteful and rife with fraud. His views on humanitarian assistance have seesawed since he entered political life."                                                            
#> [22] " By Andrew JacobsSaurabh Datar and Antonio de Luca"                                                                                                                                                                                                                                    
#> [23] "Promise of Victory Over H.I.V. Fades as U.S. Withdraws Support"                                                                                                                                                                                                                        
#> [24] "A new drug that gives almost complete protection against the virus was to be administered across Africa this year. Now, much of the funding for that effort is gone."                                                                                                                  
#> [25] " By Stephanie Nolen"                                                                                                                                                                                                                                                                   
#> [26] "Advertisement"                                                                                                                                                                                                                                                                         
#> [27] "The C.D.C.’s Vaccine Meeting: What to Watch For"                                                                                                                                                                                                                                       
#> [28] "Hints of a more skeptical approach to immunizations have already surfaced."                                                                                                                                                                                                            
#> [29] " By Apoorva Mandavilli"                                                                                                                                                                                                                                                                
#> [30] "‘I Feel Like I’ve Been Lied To’: When a Measles Outbreak Hits Home"                                                                                                                                                                                                                    
#> [31] "From a lone clinic in Texas to an entire school district in North Dakota, the virus is upending daily life and revealing a deeper crisis of belief."                                                                                                                                   
#> [32] " By Eli Saslow and Erin Schaff"                                                                                                                                                                                                                                                        
#> [33] "How Kennedy’s Purge of Advisers Could Disrupt U.S. Vaccinations"                                                                                                                                                                                                                       
#> [34] "Robert F. Kennedy Jr., the health secretary, dismissed 17 scientific advisers to the C.D.C. Critics fear newly appointed members will roll back vaccine recommendations."                                                                                                              
#> [35] " By Apoorva Mandavilli"                                                                                                                                                                                                                                                                
#> [36] "Tracking U.S. Measles Outbreaks"                                                                                                                                                                                                                                                       
#> [37] "As new cases are reported, our maps show the spread of the virus."                                                                                                                                                                                                                     
#> [38] " By Jonathan Corum and Teddy Rosenbluth"                                                                                                                                                                                                                                               
#> [39] "Why Is It So Hard to Stop a Measles Outbreak?"                                                                                                                                                                                                                                         
#> [40] "An 8-year-old girl recently became the second confirmed fatality in a Texas measles outbreak. Aatish Bhatia, who creates interactive articles for The Upshot, describes how a high rate of vaccination is needed to protect a community from a highly contagious disease like measles."
#> [41] " By Aatish BhatiaLaura SalaberryRebecca SunerZach Wood and Francesca Paris"                                                                                                                                                                                                            
#> [42] "Many Older People Embrace Vaccines. Research Is Proving Them Right."                                                                                                                                                                                                                   
#> [43] "Newer formulations are even more effective at preventing illnesses that commonly afflict seniors — perhaps even dementia."                                                                                                                                                             
#> [44] " By Paula Span"                                                                                                                                                                                                                                                                        
#> [45] "Health Risks From Cannabis May Grow as Users Age"                                                                                                                                                                                                                                      
#> [46] "More older people are using cannabis products regularly, but research suggests their cannabis-related health problems are also on the rise."                                                                                                                                           
#> [47] " By Paula Span"                                                                                                                                                                                                                                                                        
#> [48] "A Ministroke Can Have Major Consequences"                                                                                                                                                                                                                                              
#> [49] "So-called transient ischemic attacks can eventually lead to cognitive declines as steep as those following a full-on stroke, new research finds."                                                                                                                                      
#> [50] " By Paula Span"                                                                                                                                                                                                                                                                        
#> [51] "Honey, Sweetie, Dearie: The Perils of Elderspeak"                                                                                                                                                                                                                                      
#> [52] "A new training program teaches aides to stop baby talk and address older people as adults."                                                                                                                                                                                            
#> [53] " By Paula Span"                                                                                                                                                                                                                                                                        
#> [54] "When They Don’t Recognize You Anymore"                                                                                                                                                                                                                                                 
#> [55] "People with dementia often forget even close family members as the disease advances. “It can throw people into an existential crisis,” one expert said."                                                                                                                               
#> [56] " By Paula Span"                                                                                                                                                                                                                                                                        
#> [57] "Advertisement"                                                                                                                                                                                                                                                                         
#> [58] "A Common Assumption About Aging May Be Wrong, Study Suggests"                                                                                                                                                                                                                          
#> [59] "Experts have long pointed to inflammation as a natural part of getting older. But a new paper suggests it might be more a product of our environment."                                                                                                                                 
#> [60] " By Mohana Ravindranath"                                                                                                                                                                                                                                                               
#> [61] "7 Exercises to Soothe Your Sore Neck and Shoulders"                                                                                                                                                                                                                                    
#> [62] "Try this simple routine to combat “tech neck.”"                                                                                                                                                                                                                                        
#> [63] " By Anna Maltby and Theodore Tae"                                                                                                                                                                                                                                                      
#> [64] "How to Do a Perfect Lunge"                                                                                                                                                                                                                                                             
#> [65] "Mastering the lower-body move can improve your balance and stability."                                                                                                                                                                                                                 
#> [66] " By Anna Maltby"                                                                                                                                                                                                                                                                       
#> [67] "Happiness Doesn’t Have to Be a Heavy Lift"                                                                                                                                                                                                                                             
#> [68] "It only takes a few minutes to boost your mood."                                                                                                                                                                                                                                       
#> [69] " By Jancee Dunn"                                                                                                                                                                                                                                                                       
#> [70] "What to Know About the Vaccine Preservative Thimerosal"                                                                                                                                                                                                                                
#> [71] "A C.D.C. advisory committee on Thursday voted against flu shots that contain the ingredient."                                                                                                                                                                                          
#> [72] " By Teddy Rosenbluth"                                                                                                                                                                                                                                                                  
#> [73] "Robert F. Kennedy Jr., the health and human services secretary, said the United States would not deliver on a $1.2 billion pledge made by the Biden administration until the organization changed its processes."                                                                      
#> [74] "By The New York Times"                                                                                                                                                                                                                                                                 
#> [75] "The Senate health committee chairman said new members of a key advisory panel who were appointed by Robert F. Kennedy Jr. “lack experience.”"                                                                                                                                          
#> [76] "By Sheryl Gay Stolberg"                                                                                                                                                                                                                                                                
#> [77] "Temperatures in Central Park this week could reach 100 degrees for the first time since 2012."                                                                                                                                                                                         
#> [78] "By Amy Graff"                                                                                                                                                                                                                                                                          
#> [79] "Most in a small group of patients receiving a stem cell-based infusion no longer needed insulin, but the drug may not suit those with more manageable type 1 diabetes."                                                                                                                
#> [80] "By Gina Kolata"                                                                                                                                                                                                                                                                        
#> [81] "Major companies had faced mounting pressure to stop denying or stalling authorization of coverage for treatments and prescriptions."                                                                                                                                                   
#> [82] "By Reed Abelson"                                                                                                                                                                                                                                                                       
#> [83] "Proponents say that manually stimulating acupressure points can ease a variety of maladies."                                                                                                                                                                                           
#> [84] "By Christina Caron"                                                                                                                                                                                                                                                                    
#> [85] "Once nearly eradicated, the “old man’s disease” is back and suffocating younger miners. Federal cuts risk putting a solution further out of reach."                                                                                                                                    
#> [86] "By Kate Morgan and Jared Hamilton"                                                                                                                                                                                                                                                     
#> [87] "Limits on travel and visa appointments have delayed or prevented foreign doctors from entering the country for jobs set to begin in weeks."                                                                                                                                            
#> [88] "By Roni Caryn Rabin"                                                                                                                                                                                                                                                                   
#> [89] "The drug could change the course of the AIDS epidemic. But the Trump administration has gutted the programs that might have paid for it in low-income countries."                                                                                                                      
#> [90] "By Apoorva Mandavilli"                                                                                                                                                                                                                                                                 
#> [91] "About 70,000 years ago in Africa, humans expanded into more extreme environments, a new study finds, setting the stage for our global migration."                                                                                                                                      
#> [92] "By Carl Zimmer"                                                                                                                                                                                                                                                                        
#> [93] "Advertisement"                                                                                                                                                                                                                                                                         
#> [94] "Advertisement"