Simple website scraping

This function is used to scrap one element from a website.

scrap(link, node, clean = FALSE, askRobot = FALSE)

Arguments

link
the link of the webpage to scrap

node
the CSS element to consider, the SelectorGadget tool is highly recommended

clean
logical. Should the function clean the extracted vector or not ? Default is FALSE.

askRobot
logical. Should the function ask the robots.txt if we're allowed or not to scrap the web page ? Default is FALSE.

Value

a character vector

Examples

# \donttest{
# Extracting imdb top 250 movie titles

link <- "https://www.imdb.com/chart/top/"
node <- ".titleColumn a"

scrap(link, node)# }

#>   [1] "The Shawshank Redemption"                                            
#>   [2] "The Godfather"                                                       
#>   [3] "The Godfather: Part II"                                              
#>   [4] "The Dark Knight"                                                     
#>   [5] "12 Angry Men"                                                        
#>   [6] "Schindler's List"                                                    
#>   [7] "The Lord of the Rings: The Return of the King"                       
#>   [8] "Pulp Fiction"                                                        
#>   [9] "Il buono, il brutto, il cattivo"                                     
#>  [10] "Fight Club"                                                          
#>  [11] "The Lord of the Rings: The Fellowship of the Ring"                   
#>  [12] "Forrest Gump"                                                        
#>  [13] "Inception"                                                           
#>  [14] "Star Wars: Episode V - The Empire Strikes Back"                      
#>  [15] "The Lord of the Rings: The Two Towers"                               
#>  [16] "The Matrix"                                                          
#>  [17] "Goodfellas"                                                          
#>  [18] "One Flew Over the Cuckoo's Nest"                                     
#>  [19] "Shichinin no samurai"                                                
#>  [20] "Se7en"                                                               
#>  [21] "Joker"                                                               
#>  [22] "Cidade de Deus"                                                      
#>  [23] "La vita è bella"                                                     
#>  [24] "The Silence of the Lambs"                                            
#>  [25] "It's a Wonderful Life"                                               
#>  [26] "Star Wars"                                                           
#>  [27] "Saving Private Ryan"                                                 
#>  [28] "Sen to Chihiro no kamikakushi"                                       
#>  [29] "The Green Mile"                                                      
#>  [30] "Interstellar"                                                        
#>  [31] "Léon"                                                                
#>  [32] "Gisaengchung"                                                        
#>  [33] "The Usual Suspects"                                                  
#>  [34] "Seppuku"                                                             
#>  [35] "The Lion King"                                                       
#>  [36] "American History X"                                                  
#>  [37] "Back to the Future"                                                  
#>  [38] "Terminator 2: Judgment Day"                                          
#>  [39] "The Pianist"                                                         
#>  [40] "Modern Times"                                                        
#>  [41] "Psycho"                                                              
#>  [42] "Gladiator"                                                           
#>  [43] "The Intouchables"                                                    
#>  [44] "City Lights"                                                         
#>  [45] "The Departed"                                                        
#>  [46] "Whiplash"                                                            
#>  [47] "Once Upon a Time in the West"                                        
#>  [48] "The Prestige"                                                        
#>  [49] "Casablanca"                                                          
#>  [50] "Hotaru no haka"                                                      
#>  [51] "Rear Window"                                                         
#>  [52] "Nuovo Cinema Paradiso"                                               
#>  [53] "Alien"                                                               
#>  [54] "Raiders of the Lost Ark"                                             
#>  [55] "Apocalypse Now"                                                      
#>  [56] "Memento"                                                             
#>  [57] "The Great Dictator"                                                  
#>  [58] "Avengers: Endgame"                                                   
#>  [59] "The Lives of Others"                                                 
#>  [60] "Django Unchained"                                                    
#>  [61] "Avengers: Infinity War"                                              
#>  [62] "The Shining"                                                         
#>  [63] "Paths of Glory"                                                      
#>  [64] "Spider-Man: Into the Spider-Verse"                                   
#>  [65] "WALL·E"                                                              
#>  [66] "Mononoke-hime"                                                       
#>  [67] "Sunset Blvd."                                                        
#>  [68] "Dr. Strangelove or: How I Learned to Stop Worrying and Love the Bomb"
#>  [69] "Oldeuboi"                                                            
#>  [70] "Witness for the Prosecution"                                         
#>  [71] "The Dark Knight Rises"                                               
#>  [72] "Once Upon a Time in America"                                         
#>  [73] "Aliens"                                                              
#>  [74] "Kimi no na wa."                                                      
#>  [75] "American Beauty"                                                     
#>  [76] "Coco"                                                                
#>  [77] "Braveheart"                                                          
#>  [78] "Das Boot"                                                            
#>  [79] "3 Idiots"                                                            
#>  [80] "Star Wars: Episode VI - Return of the Jedi"                          
#>  [81] "Taare Zameen Par"                                                    
#>  [82] "Toy Story"                                                           
#>  [83] "Amadeus"                                                             
#>  [84] "Reservoir Dogs"                                                      
#>  [85] "Inglourious Basterds"                                                
#>  [86] "Good Will Hunting"                                                   
#>  [87] "M - Eine Stadt sucht einen Mörder"                                   
#>  [88] "Requiem for a Dream"                                                 
#>  [89] "2001: A Space Odyssey"                                               
#>  [90] "Dangal"                                                              
#>  [91] "Vertigo"                                                             
#>  [92] "Eternal Sunshine of the Spotless Mind"                               
#>  [93] "Citizen Kane"                                                        
#>  [94] "Jagten"                                                              
#>  [95] "Full Metal Jacket"                                                   
#>  [96] "North by Northwest"                                                  
#>  [97] "A Clockwork Orange"                                                  
#>  [98] "Snatch"                                                              
#>  [99] "The Kid"                                                             
#> [100] "Ladri di biciclette"                                                 
#> [101] "Amélie"                                                              
#> [102] "Scarface"                                                            
#> [103] "Singin' in the Rain"                                                 
#> [104] "Taxi Driver"                                                         
#> [105] "Lawrence of Arabia"                                                  
#> [106] "Toy Story 3"                                                         
#> [107] "The Sting"                                                           
#> [108] "Metropolis"                                                          
#> [109] "Per qualche dollaro in più"                                          
#> [110] "Capharnaüm"                                                          
#> [111] "Ikiru"                                                               
#> [112] "Double Indemnity"                                                    
#> [113] "Jodaeiye Nader az Simin"                                             
#> [114] "To Kill a Mockingbird"                                               
#> [115] "The Apartment"                                                       
#> [116] "Indiana Jones and the Last Crusade"                                  
#> [117] "Up"                                                                  
#> [118] "Incendies"                                                           
#> [119] "L.A. Confidential"                                                   
#> [120] "Monty Python and the Holy Grail"                                     
#> [121] "Heat"                                                                
#> [122] "Rashômon"                                                            
#> [123] "Die Hard"                                                            
#> [124] "Batman Begins"                                                       
#> [125] "Yôjinbô"                                                             
#> [126] "Unforgiven"                                                          
#> [127] "Green Book"                                                          
#> [128] "Ford v Ferrari"                                                      
#> [129] "Der Untergang"                                                       
#> [130] "Bacheha-Ye aseman"                                                   
#> [131] "Some Like It Hot"                                                    
#> [132] "Hauru no ugoku shiro"                                                
#> [133] "The Great Escape"                                                    
#> [134] "Tonari no Totoro"                                                    
#> [135] "Ran"                                                                 
#> [136] "A Beautiful Mind"                                                    
#> [137] "All About Eve"                                                       
#> [138] "Casino"                                                              
#> [139] "Pan's Labyrinth"                                                     
#> [140] "Idi i smotri"                                                        
#> [141] "Raging Bull"                                                         
#> [142] "El secreto de sus ojos"                                              
#> [143] "Lock, Stock and Two Smoking Barrels"                                 
#> [144] "The Wolf of Wall Street"                                             
#> [145] "The Treasure of the Sierra Madre"                                    
#> [146] "Judgment at Nuremberg"                                               
#> [147] "Three Billboards Outside Ebbing, Missouri"                           
#> [148] "Babam ve Oglum"                                                      
#> [149] "Chinatown"                                                           
#> [150] "Marriage Story"                                                      
#> [151] "The Gold Rush"                                                       
#> [152] "There Will Be Blood"                                                 
#> [153] "Dial M for Murder"                                                   
#> [154] "V for Vendetta"                                                      
#> [155] "Inside Out"                                                          
#> [156] "Det sjunde inseglet"                                                 
#> [157] "Warrior"                                                             
#> [158] "No Country for Old Men"                                              
#> [159] "Trainspotting"                                                       
#> [160] "The Elephant Man"                                                    
#> [161] "Room"                                                                
#> [162] "Shutter Island"                                                      
#> [163] "The Sixth Sense"                                                     
#> [164] "The Thing"                                                           
#> [165] "Gone with the Wind"                                                  
#> [166] "The Bridge on the River Kwai"                                        
#> [167] "Blade Runner"                                                        
#> [168] "The Third Man"                                                       
#> [169] "Jurassic Park"                                                       
#> [170] "On the Waterfront"                                                   
#> [171] "Finding Nemo"                                                        
#> [172] "Smultronstället"                                                     
#> [173] "Andhadhun"                                                           
#> [174] "Fargo"                                                               
#> [175] "Gran Torino"                                                         
#> [176] "Kill Bill: Vol. 1"                                                   
#> [177] "Eskiya"                                                              
#> [178] "The Deer Hunter"                                                     
#> [179] "Tôkyô monogatari"                                                    
#> [180] "Stalker"                                                             
#> [181] "The Truman Show"                                                     
#> [182] "Drishyam"                                                            
#> [183] "The Big Lebowski"                                                    
#> [184] "Relatos salvajes"                                                    
#> [185] "Mary and Max"                                                        
#> [186] "The Irishman"                                                        
#> [187] "In the Name of the Father"                                           
#> [188] "Gone Girl"                                                           
#> [189] "Klaus"                                                               
#> [190] "Hacksaw Ridge"                                                       
#> [191] "Salinui chueok"                                                      
#> [192] "Mr. Smith Goes to Washington"                                        
#> [193] "How to Train Your Dragon"                                            
#> [194] "The General"                                                         
#> [195] "The Grand Budapest Hotel"                                            
#> [196] "Sherlock Jr."                                                        
#> [197] "Persona"                                                             
#> [198] "Before Sunrise"                                                      
#> [199] "Catch Me If You Can"                                                 
#> [200] "Cool Hand Luke"                                                      
#> [201] "Prisoners"                                                           
#> [202] "Into the Wild"                                                       
#> [203] "12 Years a Slave"                                                    
#> [204] "Le salaire de la peur"                                               
#> [205] "Network"                                                             
#> [206] "Stand by Me"                                                         
#> [207] "Life of Brian"                                                       
#> [208] "Mad Max: Fury Road"                                                  
#> [209] "Platoon"                                                             
#> [210] "La passion de Jeanne d'Arc"                                          
#> [211] "Barry Lyndon"                                                        
#> [212] "Million Dollar Baby"                                                 
#> [213] "Hachi: A Dog's Tale"                                                 
#> [214] "Rush"                                                                
#> [215] "Rang De Basanti"                                                     
#> [216] "Logan"                                                               
#> [217] "Andrei Rublev"                                                       
#> [218] "Ben-Hur"                                                             
#> [219] "Kaze no tani no Naushika"                                            
#> [220] "Les quatre cents coups"                                              
#> [221] "Harry Potter and the Deathly Hallows: Part 2"                        
#> [222] "Hotel Rwanda"                                                        
#> [223] "The Red Shoes"                                                       
#> [224] "Amores perros"                                                       
#> [225] "Dead Poets Society"                                                  
#> [226] "Spotlight"                                                           
#> [227] "Rebecca"                                                             
#> [228] "Ace in the Hole"                                                     
#> [229] "Rocky"                                                               
#> [230] "White Heat"                                                          
#> [231] "Ah-ga-ssi"                                                           
#> [232] "Monsters, Inc."                                                      
#> [233] "It Happened One Night"                                               
#> [234] "La haine"                                                            
#> [235] "Lagaan: Once Upon a Time in India"                                   
#> [236] "Gangs of Wasseypur"                                                  
#> [237] "The Princess Bride"                                                  
#> [238] "Before Sunset"                                                       
#> [239] "The Help"                                                            
#> [240] "PK"                                                                  
#> [241] "Butch Cassidy and the Sundance Kid"                                  
#> [242] "Faa yeung nin wa"                                                    
#> [243] "Knives Out"                                                          
#> [244] "The Terminator"                                                      
#> [245] "Akira"                                                               
#> [246] "Paris, Texas"                                                        
#> [247] "Aladdin"                                                             
#> [248] "Munna Bhai M.B.B.S."                                                 
#> [249] "La battaglia di Algeri"                                              
#> [250] "Guardians of the Galaxy"