Simple website scraping

This function is used to scrape one element from a website.

scrap(link, node, clean = FALSE, askRobot = FALSE)

Arguments

link
the link of the web page to scrape

node
the HTML or CSS element to consider, the SelectorGadget tool is highly recommended

clean
logical. Should the function clean the extracted vector or not ? Default is FALSE.

askRobot
logical. Should the function ask the robots.txt if we're allowed or not to scrape the web page ? Default is FALSE.

Value

a character vector

Examples

# \donttest{
# Extracting imdb top 250 movie titles

link <- "https://www.imdb.com/chart/top/"
node <- ".titleColumn a"

scrap(link, node)# }

#>   [1] "The Shawshank Redemption"                                            
#>   [2] "The Godfather"                                                       
#>   [3] "The Godfather: Part II"                                              
#>   [4] "The Dark Knight"                                                     
#>   [5] "12 Angry Men"                                                        
#>   [6] "Schindler's List"                                                    
#>   [7] "The Lord of the Rings: The Return of the King"                       
#>   [8] "Pulp Fiction"                                                        
#>   [9] "Il buono, il brutto, il cattivo"                                     
#>  [10] "The Lord of the Rings: The Fellowship of the Ring"                   
#>  [11] "Fight Club"                                                          
#>  [12] "Forrest Gump"                                                        
#>  [13] "Inception"                                                           
#>  [14] "The Lord of the Rings: The Two Towers"                               
#>  [15] "Star Wars: Episode V - The Empire Strikes Back"                      
#>  [16] "The Matrix"                                                          
#>  [17] "Goodfellas"                                                          
#>  [18] "One Flew Over the Cuckoo's Nest"                                     
#>  [19] "Shichinin no samurai"                                                
#>  [20] "Se7en"                                                               
#>  [21] "La vita è bella"                                                     
#>  [22] "Cidade de Deus"                                                      
#>  [23] "The Silence of the Lambs"                                            
#>  [24] "It's a Wonderful Life"                                               
#>  [25] "Star Wars"                                                           
#>  [26] "Saving Private Ryan"                                                 
#>  [27] "Sen to Chihiro no kamikakushi"                                       
#>  [28] "The Green Mile"                                                      
#>  [29] "Interstellar"                                                        
#>  [30] "Gisaengchung"                                                        
#>  [31] "Léon"                                                                
#>  [32] "The Usual Suspects"                                                  
#>  [33] "Seppuku"                                                             
#>  [34] "The Lion King"                                                       
#>  [35] "Back to the Future"                                                  
#>  [36] "The Pianist"                                                         
#>  [37] "Terminator 2: Judgment Day"                                          
#>  [38] "American History X"                                                  
#>  [39] "Modern Times"                                                        
#>  [40] "Psycho"                                                              
#>  [41] "Gladiator"                                                           
#>  [42] "City Lights"                                                         
#>  [43] "The Departed"                                                        
#>  [44] "The Intouchables"                                                    
#>  [45] "Whiplash"                                                            
#>  [46] "The Prestige"                                                        
#>  [47] "Hamilton"                                                            
#>  [48] "Hotaru no haka"                                                      
#>  [49] "Once Upon a Time in the West"                                        
#>  [50] "Casablanca"                                                          
#>  [51] "Nuovo Cinema Paradiso"                                               
#>  [52] "Rear Window"                                                         
#>  [53] "Alien"                                                               
#>  [54] "Apocalypse Now"                                                      
#>  [55] "Memento"                                                             
#>  [56] "The Great Dictator"                                                  
#>  [57] "Raiders of the Lost Ark"                                             
#>  [58] "Django Unchained"                                                    
#>  [59] "The Lives of Others"                                                 
#>  [60] "Joker"                                                               
#>  [61] "Paths of Glory"                                                      
#>  [62] "WALL·E"                                                              
#>  [63] "The Shining"                                                         
#>  [64] "Avengers: Infinity War"                                              
#>  [65] "Sunset Blvd."                                                        
#>  [66] "Witness for the Prosecution"                                         
#>  [67] "Mononoke-hime"                                                       
#>  [68] "Oldeuboi"                                                            
#>  [69] "Spider-Man: Into the Spider-Verse"                                   
#>  [70] "Dr. Strangelove or: How I Learned to Stop Worrying and Love the Bomb"
#>  [71] "The Dark Knight Rises"                                               
#>  [72] "Once Upon a Time in America"                                         
#>  [73] "Aliens"                                                              
#>  [74] "Kimi no na wa."                                                      
#>  [75] "Avengers: Endgame"                                                   
#>  [76] "Coco"                                                                
#>  [77] "American Beauty"                                                     
#>  [78] "Braveheart"                                                          
#>  [79] "3 Idiots"                                                            
#>  [80] "Das Boot"                                                            
#>  [81] "Toy Story"                                                           
#>  [82] "Tengoku to jigoku"                                                   
#>  [83] "Amadeus"                                                             
#>  [84] "Capharnaüm"                                                          
#>  [85] "Inglourious Basterds"                                                
#>  [86] "Star Wars: Episode VI - Return of the Jedi"                          
#>  [87] "Taare Zameen Par"                                                    
#>  [88] "Good Will Hunting"                                                   
#>  [89] "Reservoir Dogs"                                                      
#>  [90] "2001: A Space Odyssey"                                               
#>  [91] "Requiem for a Dream"                                                 
#>  [92] "Vertigo"                                                             
#>  [93] "M - Eine Stadt sucht einen Mörder"                                   
#>  [94] "Eternal Sunshine of the Spotless Mind"                               
#>  [95] "Dangal"                                                              
#>  [96] "Jagten"                                                              
#>  [97] "Citizen Kane"                                                        
#>  [98] "1917"                                                                
#>  [99] "Full Metal Jacket"                                                   
#> [100] "Ladri di biciclette"                                                 
#> [101] "The Kid"                                                             
#> [102] "Singin' in the Rain"                                                 
#> [103] "A Clockwork Orange"                                                  
#> [104] "North by Northwest"                                                  
#> [105] "Snatch"                                                              
#> [106] "Scarface"                                                            
#> [107] "Taxi Driver"                                                         
#> [108] "Ikiru"                                                               
#> [109] "Amélie"                                                              
#> [110] "Lawrence of Arabia"                                                  
#> [111] "Toy Story 3"                                                         
#> [112] "The Sting"                                                           
#> [113] "Metropolis"                                                          
#> [114] "Jodaeiye Nader az Simin"                                             
#> [115] "Incendies"                                                           
#> [116] "Per qualche dollaro in più"                                          
#> [117] "Idi i smotri"                                                        
#> [118] "The Apartment"                                                       
#> [119] "Double Indemnity"                                                    
#> [120] "To Kill a Mockingbird"                                               
#> [121] "Up"                                                                  
#> [122] "Indiana Jones and the Last Crusade"                                  
#> [123] "Heat"                                                                
#> [124] "L.A. Confidential"                                                   
#> [125] "Die Hard"                                                            
#> [126] "Monty Python and the Holy Grail"                                     
#> [127] "Rashômon"                                                            
#> [128] "Yôjinbô"                                                             
#> [129] "Batman Begins"                                                       
#> [130] "Green Book"                                                          
#> [131] "Der Untergang"                                                       
#> [132] "Bacheha-Ye aseman"                                                   
#> [133] "Unforgiven"                                                          
#> [134] "Ran"                                                                 
#> [135] "Some Like It Hot"                                                    
#> [136] "Hauru no ugoku shiro"                                                
#> [137] "A Beautiful Mind"                                                    
#> [138] "All About Eve"                                                       
#> [139] "Casino"                                                              
#> [140] "The Great Escape"                                                    
#> [141] "The Wolf of Wall Street"                                             
#> [142] "Pan's Labyrinth"                                                     
#> [143] "El secreto de sus ojos"                                              
#> [144] "Lock, Stock and Two Smoking Barrels"                                 
#> [145] "Anand"                                                               
#> [146] "There Will Be Blood"                                                 
#> [147] "Raging Bull"                                                         
#> [148] "Tonari no Totoro"                                                    
#> [149] "Judgment at Nuremberg"                                               
#> [150] "The Treasure of the Sierra Madre"                                    
#> [151] "Dial M for Murder"                                                   
#> [152] "Three Billboards Outside Ebbing, Missouri"                           
#> [153] "Chinatown"                                                           
#> [154] "The Gold Rush"                                                       
#> [155] "Babam ve Oglum"                                                      
#> [156] "Shutter Island"                                                      
#> [157] "No Country for Old Men"                                              
#> [158] "V for Vendetta"                                                      
#> [159] "Det sjunde inseglet"                                                 
#> [160] "Inside Out"                                                          
#> [161] "Warrior"                                                             
#> [162] "The Elephant Man"                                                    
#> [163] "The Thing"                                                           
#> [164] "The Sixth Sense"                                                     
#> [165] "Trainspotting"                                                       
#> [166] "Jurassic Park"                                                       
#> [167] "Gone with the Wind"                                                  
#> [168] "The Truman Show"                                                     
#> [169] "Finding Nemo"                                                        
#> [170] "Smultronstället"                                                     
#> [171] "Blade Runner"                                                        
#> [172] "Stalker"                                                             
#> [173] "Kill Bill: Vol. 1"                                                   
#> [174] "The Bridge on the River Kwai"                                        
#> [175] "Room"                                                                
#> [176] "Fargo"                                                               
#> [177] "Salinui chueok"                                                      
#> [178] "Tôkyô monogatari"                                                    
#> [179] "The Third Man"                                                       
#> [180] "Gran Torino"                                                         
#> [181] "On the Waterfront"                                                   
#> [182] "Vikram Vedha"                                                        
#> [183] "Relatos salvajes"                                                    
#> [184] "The Deer Hunter"                                                     
#> [185] "Klaus"                                                               
#> [186] "In the Name of the Father"                                           
#> [187] "Mary and Max"                                                        
#> [188] "Gone Girl"                                                           
#> [189] "The Grand Budapest Hotel"                                            
#> [190] "Hacksaw Ridge"                                                       
#> [191] "Before Sunrise"                                                      
#> [192] "Andhadhun"                                                           
#> [193] "Catch Me If You Can"                                                 
#> [194] "The Big Lebowski"                                                    
#> [195] "Persona"                                                             
#> [196] "To Be or Not to Be"                                                  
#> [197] "Prisoners"                                                           
#> [198] "Sherlock Jr."                                                        
#> [199] "Eskiya"                                                              
#> [200] "The General"                                                         
#> [201] "How to Train Your Dragon"                                            
#> [202] "Ford v Ferrari"                                                      
#> [203] "Mr. Smith Goes to Washington"                                        
#> [204] "12 Years a Slave"                                                    
#> [205] "Barry Lyndon"                                                        
#> [206] "Mad Max: Fury Road"                                                  
#> [207] "Million Dollar Baby"                                                 
#> [208] "Stand by Me"                                                         
#> [209] "Network"                                                             
#> [210] "Cool Hand Luke"                                                      
#> [211] "Dead Poets Society"                                                  
#> [212] "Ben-Hur"                                                             
#> [213] "Hachi: A Dog's Tale"                                                 
#> [214] "Harry Potter and the Deathly Hallows: Part 2"                        
#> [215] "Platoon"                                                             
#> [216] "Into the Wild"                                                       
#> [217] "Logan"                                                               
#> [218] "Le salaire de la peur"                                               
#> [219] "Life of Brian"                                                       
#> [220] "Rush"                                                                
#> [221] "Ah-ga-ssi"                                                           
#> [222] "Les quatre cents coups"                                              
#> [223] "La passion de Jeanne d'Arc"                                          
#> [224] "Andrei Rublev"                                                       
#> [225] "Hotel Rwanda"                                                        
#> [226] "Spotlight"                                                           
#> [227] "Amores perros"                                                       
#> [228] "La haine"                                                            
#> [229] "Du rififi chez les hommes"                                           
#> [230] "Kaze no tani no Naushika"                                            
#> [231] "Rocky"                                                               
#> [232] "Gangs of Wasseypur"                                                  
#> [233] "Monsters, Inc."                                                      
#> [234] "Rebecca"                                                             
#> [235] "Rang De Basanti"                                                     
#> [236] "Before Sunset"                                                       
#> [237] "Portrait de la jeune fille en feu"                                   
#> [238] "Fa yeung nin wah"                                                    
#> [239] "Paris, Texas"                                                        
#> [240] "It Happened One Night"                                               
#> [241] "Drishyam"                                                            
#> [242] "Contratiempo"                                                        
#> [243] "The Princess Bride"                                                  
#> [244] "The Help"                                                            
#> [245] "The Circus"                                                          
#> [246] "La battaglia di Algeri"                                              
#> [247] "The Terminator"                                                      
#> [248] "Koe no katachi"                                                      
#> [249] "Aladdin"                                                             
#> [250] "Kis Uykusu"