Simple website scraping

This function is used to scrap one element from a website.

scrap(link, node, clean = FALSE, askRobot = FALSE)

Arguments

link
the link of the webpage to scrap

node
the CSS element to consider, the SelectorGadget tool is highly recommended

clean
logical. Should the function clean the extracted vector or not ? Default is FALSE.

askRobot
logical. Should the function ask the robots.txt if we're allowed or not to scrap the web page ? Default is FALSE.

Value

a character vector

Examples

# \donttest{
# Extracting imdb top 250 movie titles

link <- "https://www.imdb.com/chart/top/"
node <- ".titleColumn a"

scrap(link, node)# }

#>   [1] "The Shawshank Redemption"                                            
#>   [2] "The Godfather"                                                       
#>   [3] "The Godfather: Part II"                                              
#>   [4] "The Dark Knight"                                                     
#>   [5] "12 Angry Men"                                                        
#>   [6] "Schindler's List"                                                    
#>   [7] "The Lord of the Rings: The Return of the King"                       
#>   [8] "Pulp Fiction"                                                        
#>   [9] "Il buono, il brutto, il cattivo"                                     
#>  [10] "The Lord of the Rings: The Fellowship of the Ring"                   
#>  [11] "Fight Club"                                                          
#>  [12] "Forrest Gump"                                                        
#>  [13] "Inception"                                                           
#>  [14] "Star Wars: Episode V - The Empire Strikes Back"                      
#>  [15] "The Lord of the Rings: The Two Towers"                               
#>  [16] "The Matrix"                                                          
#>  [17] "Goodfellas"                                                          
#>  [18] "One Flew Over the Cuckoo's Nest"                                     
#>  [19] "Shichinin no samurai"                                                
#>  [20] "Se7en"                                                               
#>  [21] "Cidade de Deus"                                                      
#>  [22] "La vita è bella"                                                     
#>  [23] "The Silence of the Lambs"                                            
#>  [24] "It's a Wonderful Life"                                               
#>  [25] "Star Wars"                                                           
#>  [26] "Saving Private Ryan"                                                 
#>  [27] "Gisaengchung"                                                        
#>  [28] "Sen to Chihiro no kamikakushi"                                       
#>  [29] "Joker"                                                               
#>  [30] "The Green Mile"                                                      
#>  [31] "Interstellar"                                                        
#>  [32] "Léon"                                                                
#>  [33] "Seppuku"                                                             
#>  [34] "The Usual Suspects"                                                  
#>  [35] "The Lion King"                                                       
#>  [36] "American History X"                                                  
#>  [37] "Back to the Future"                                                  
#>  [38] "Terminator 2: Judgment Day"                                          
#>  [39] "The Pianist"                                                         
#>  [40] "Modern Times"                                                        
#>  [41] "Psycho"                                                              
#>  [42] "Gladiator"                                                           
#>  [43] "City Lights"                                                         
#>  [44] "The Intouchables"                                                    
#>  [45] "The Departed"                                                        
#>  [46] "Whiplash"                                                            
#>  [47] "Once Upon a Time in the West"                                        
#>  [48] "The Prestige"                                                        
#>  [49] "Casablanca"                                                          
#>  [50] "Hotaru no haka"                                                      
#>  [51] "Rear Window"                                                         
#>  [52] "Nuovo Cinema Paradiso"                                               
#>  [53] "Alien"                                                               
#>  [54] "Raiders of the Lost Ark"                                             
#>  [55] "Apocalypse Now"                                                      
#>  [56] "Memento"                                                             
#>  [57] "The Great Dictator"                                                  
#>  [58] "The Lives of Others"                                                 
#>  [59] "Django Unchained"                                                    
#>  [60] "Avengers: Endgame"                                                   
#>  [61] "Avengers: Infinity War"                                              
#>  [62] "The Shining"                                                         
#>  [63] "Paths of Glory"                                                      
#>  [64] "WALL·E"                                                              
#>  [65] "Spider-Man: Into the Spider-Verse"                                   
#>  [66] "Sunset Blvd."                                                        
#>  [67] "Mononoke-hime"                                                       
#>  [68] "Dr. Strangelove or: How I Learned to Stop Worrying and Love the Bomb"
#>  [69] "Oldeuboi"                                                            
#>  [70] "Witness for the Prosecution"                                         
#>  [71] "The Dark Knight Rises"                                               
#>  [72] "Once Upon a Time in America"                                         
#>  [73] "Aliens"                                                              
#>  [74] "Kimi no na wa."                                                      
#>  [75] "American Beauty"                                                     
#>  [76] "Coco"                                                                
#>  [77] "Braveheart"                                                          
#>  [78] "Das Boot"                                                            
#>  [79] "3 Idiots"                                                            
#>  [80] "Star Wars: Episode VI - Return of the Jedi"                          
#>  [81] "Taare Zameen Par"                                                    
#>  [82] "Toy Story"                                                           
#>  [83] "Amadeus"                                                             
#>  [84] "Reservoir Dogs"                                                      
#>  [85] "Inglourious Basterds"                                                
#>  [86] "Good Will Hunting"                                                   
#>  [87] "2001: A Space Odyssey"                                               
#>  [88] "Requiem for a Dream"                                                 
#>  [89] "M - Eine Stadt sucht einen Mörder"                                   
#>  [90] "Vertigo"                                                             
#>  [91] "Dangal"                                                              
#>  [92] "Eternal Sunshine of the Spotless Mind"                               
#>  [93] "Citizen Kane"                                                        
#>  [94] "Jagten"                                                              
#>  [95] "Full Metal Jacket"                                                   
#>  [96] "North by Northwest"                                                  
#>  [97] "A Clockwork Orange"                                                  
#>  [98] "Snatch"                                                              
#>  [99] "The Kid"                                                             
#> [100] "Ladri di biciclette"                                                 
#> [101] "Scarface"                                                            
#> [102] "Amélie"                                                              
#> [103] "Singin' in the Rain"                                                 
#> [104] "Taxi Driver"                                                         
#> [105] "Lawrence of Arabia"                                                  
#> [106] "Toy Story 3"                                                         
#> [107] "The Sting"                                                           
#> [108] "Metropolis"                                                          
#> [109] "Per qualche dollaro in più"                                          
#> [110] "Capharnaüm"                                                          
#> [111] "Jodaeiye Nader az Simin"                                             
#> [112] "Double Indemnity"                                                    
#> [113] "Ikiru"                                                               
#> [114] "To Kill a Mockingbird"                                               
#> [115] "The Apartment"                                                       
#> [116] "Indiana Jones and the Last Crusade"                                  
#> [117] "Up"                                                                  
#> [118] "Incendies"                                                           
#> [119] "L.A. Confidential"                                                   
#> [120] "Monty Python and the Holy Grail"                                     
#> [121] "Heat"                                                                
#> [122] "Die Hard"                                                            
#> [123] "Rashômon"                                                            
#> [124] "Yôjinbô"                                                             
#> [125] "Batman Begins"                                                       
#> [126] "Unforgiven"                                                          
#> [127] "Green Book"                                                          
#> [128] "Der Untergang"                                                       
#> [129] "Bacheha-Ye aseman"                                                   
#> [130] "Some Like It Hot"                                                    
#> [131] "Ford v Ferrari"                                                      
#> [132] "Hauru no ugoku shiro"                                                
#> [133] "Tonari no Totoro"                                                    
#> [134] "The Great Escape"                                                    
#> [135] "Ran"                                                                 
#> [136] "A Beautiful Mind"                                                    
#> [137] "All About Eve"                                                       
#> [138] "Idi i smotri"                                                        
#> [139] "Casino"                                                              
#> [140] "Pan's Labyrinth"                                                     
#> [141] "Raging Bull"                                                         
#> [142] "El secreto de sus ojos"                                              
#> [143] "Lock, Stock and Two Smoking Barrels"                                 
#> [144] "The Wolf of Wall Street"                                             
#> [145] "The Treasure of the Sierra Madre"                                    
#> [146] "Judgment at Nuremberg"                                               
#> [147] "Three Billboards Outside Ebbing, Missouri"                           
#> [148] "Chinatown"                                                           
#> [149] "Babam ve Oglum"                                                      
#> [150] "The Gold Rush"                                                       
#> [151] "There Will Be Blood"                                                 
#> [152] "Dial M for Murder"                                                   
#> [153] "V for Vendetta"                                                      
#> [154] "Inside Out"                                                          
#> [155] "Det sjunde inseglet"                                                 
#> [156] "Warrior"                                                             
#> [157] "No Country for Old Men"                                              
#> [158] "Trainspotting"                                                       
#> [159] "The Elephant Man"                                                    
#> [160] "Room"                                                                
#> [161] "Shutter Island"                                                      
#> [162] "The Sixth Sense"                                                     
#> [163] "The Thing"                                                           
#> [164] "Gone with the Wind"                                                  
#> [165] "Blade Runner"                                                        
#> [166] "The Bridge on the River Kwai"                                        
#> [167] "Jurassic Park"                                                       
#> [168] "The Third Man"                                                       
#> [169] "On the Waterfront"                                                   
#> [170] "Finding Nemo"                                                        
#> [171] "Smultronstället"                                                     
#> [172] "Klaus"                                                               
#> [173] "Fargo"                                                               
#> [174] "Andhadhun"                                                           
#> [175] "Gran Torino"                                                         
#> [176] "Kill Bill: Vol. 1"                                                   
#> [177] "The Deer Hunter"                                                     
#> [178] "Eskiya"                                                              
#> [179] "The Truman Show"                                                     
#> [180] "Tôkyô monogatari"                                                    
#> [181] "Stalker"                                                             
#> [182] "Relatos salvajes"                                                    
#> [183] "The Big Lebowski"                                                    
#> [184] "Mary and Max"                                                        
#> [185] "Marriage Story"                                                      
#> [186] "In the Name of the Father"                                           
#> [187] "Gone Girl"                                                           
#> [188] "Salinui chueok"                                                      
#> [189] "Hacksaw Ridge"                                                       
#> [190] "Mr. Smith Goes to Washington"                                        
#> [191] "How to Train Your Dragon"                                            
#> [192] "The General"                                                         
#> [193] "The Grand Budapest Hotel"                                            
#> [194] "Sherlock Jr."                                                        
#> [195] "Persona"                                                             
#> [196] "Before Sunrise"                                                      
#> [197] "Catch Me If You Can"                                                 
#> [198] "Prisoners"                                                           
#> [199] "Cool Hand Luke"                                                      
#> [200] "12 Years a Slave"                                                    
#> [201] "Into the Wild"                                                       
#> [202] "Le salaire de la peur"                                               
#> [203] "Network"                                                             
#> [204] "Stand by Me"                                                         
#> [205] "Life of Brian"                                                       
#> [206] "Mad Max: Fury Road"                                                  
#> [207] "Platoon"                                                             
#> [208] "Barry Lyndon"                                                        
#> [209] "La passion de Jeanne d'Arc"                                          
#> [210] "Million Dollar Baby"                                                 
#> [211] "Hachi: A Dog's Tale"                                                 
#> [212] "Rush"                                                                
#> [213] "Logan"                                                               
#> [214] "Ben-Hur"                                                             
#> [215] "Rang De Basanti"                                                     
#> [216] "Andrei Rublev"                                                       
#> [217] "Kaze no tani no Naushika"                                            
#> [218] "Harry Potter and the Deathly Hallows: Part 2"                        
#> [219] "Les quatre cents coups"                                              
#> [220] "Hotel Rwanda"                                                        
#> [221] "Dead Poets Society"                                                  
#> [222] "Amores perros"                                                       
#> [223] "Spotlight"                                                           
#> [224] "The Red Shoes"                                                       
#> [225] "Rebecca"                                                             
#> [226] "Rocky"                                                               
#> [227] "Ah-ga-ssi"                                                           
#> [228] "Ace in the Hole"                                                     
#> [229] "Monsters, Inc."                                                      
#> [230] "White Heat"                                                          
#> [231] "It Happened One Night"                                               
#> [232] "La haine"                                                            
#> [233] "The Irishman"                                                        
#> [234] "Lagaan: Once Upon a Time in India"                                   
#> [235] "The Princess Bride"                                                  
#> [236] "Gangs of Wasseypur"                                                  
#> [237] "Drishyam"                                                            
#> [238] "Before Sunset"                                                       
#> [239] "The Help"                                                            
#> [240] "Faa yeung nin wa"                                                    
#> [241] "Butch Cassidy and the Sundance Kid"                                  
#> [242] "PK"                                                                  
#> [243] "The Terminator"                                                      
#> [244] "Akira"                                                               
#> [245] "Paris, Texas"                                                        
#> [246] "Aladdin"                                                             
#> [247] "La battaglia di Algeri"                                              
#> [248] "Contratiempo"                                                        
#> [249] "Tenkû no shiro Rapyuta"                                              
#> [250] "Guardians of the Galaxy"