Website title scraping

This function is used to scrape titles (h1, h2 & h3 html tags) from a website. Useful for scraping daily electronic newspapers' titles.

titles_scrap(link, contain = NULL, case_sensitive = FALSE, askRobot = FALSE)

Arguments

link
the link of the web page to scrape

contain
filter the titles according to a character string provided.

case_sensitive
logical. Should the contain argument be case sensitive ? defaults to FALSE

askRobot
logical. Should the function ask the robots.txt if we're allowed or not to scrape the web page ? Default is FALSE

Value

a character vector

Examples

# \donttest{
# Extracting the current titles of the New York Times

link     <- "https://www.nytimes.com/"

titles_scrap(link)# }

#>  [1] "Listen to ‘The Daily’"                                                                                                   
#>  [2] "Listen to ‘The Argument’"                                                                                                
#>  [3] "In the ‘At Home’ Newsletter"                                                                                             
#>  [4] "Indonesian Jetliner Crashes Into the Sea After Takeoff"                                                                  
#>  [5] "He Dreamed of Being a Police Officer, Then Was Killed by a Pro-Trump Mob"                                                
#>  [6] "11 Journalists on Covering the Capitol Siege: ‘This Could Get Ugly’"                                                     
#>  [7] "Trump’s Legacy: Voters Who Reject Democracy and Any Politics but Their Own"                                              
#>  [8] "Bravery or reputation management? The resignations of some Trump officials are drawing skepticism."                      
#>  [9] "Here are the Trump aides who plan to stay to the end."                                                                   
#> [10] "As Coronavirus Mutates, the World Stumbles Again to Respond"                                                             
#> [11] "More Than 150,000 Are Fully Vaccinated in the U.S."                                                                      
#> [12] "False Reports of a New ‘U.S. Variant’ Came from White House Task Force"                                                  
#> [13] "Four Reasons the N.F.L. Shattered Its Scoring Record in 2020"                                                            
#> [14] "Covid-19 is forcing N.F.L. players and other pro athletes to make unusually hard decisions about work-life balance."     
#> [15] "The Weekender: The Last Two Northern White Rhinos"                                                                       
#> [16] "Did you follow the headlines this week? Take our quiz to find out."                                                      
#> [17] "Awe and Shock"                                                                                                           
#> [18] "Can Donald Trump Survive Without Twitter?"                                                                               
#> [19] "Far-Right Protesters Stormed Germany’s Parliament. What Can America Learn?"                                              
#> [20] "Listen to ‘Sway’: If You Were on Parler, You Saw the Mob Coming"                                                         
#> [21] "Impeach Now. Running Out the Clock on Trump Is Cowardly and Dangerous."                                                  
#> [22] "Stop Pretending ‘This Is Not Who We Are’"                                                                                
#> [23] "Neil Sheehan Forced an American Reckoning"                                                                               
#> [24] "Appeasement Got Us Where We Are"                                                                                         
#> [25] "This Is When the Fever Breaks"                                                                                           
#> [26] "How to Ensure This Never Happens Again"                                                                                  
#> [27] "More Immigrants Will Come to the U.S. Under President Biden. That’s a Good Thing."                                       
#> [28] "He Was Going to Close the Family Diner. Then He Got a Sign."                                                             
#> [29] "Louise Linton Has Made a Movie"                                                                                          
#> [30] "The Man Who Turned Credit-Card Points Into an Empire"                                                                    
#> [31] "Site Index"                                                                                                              
#> [32] "Site Information Navigation"                                                                                             
#> [33] "Democrats Lay Groundwork for Impeaching Trump Again"                                                                     
#> [34] "‘I Want Him Out’: Murkowski Is First G.O.P. Senator to Call for Removal"                                                 
#> [35] "Twitter Permanently Bans Trump, Capping Online Revolt"                                                                   
#> [36] "Google and Apple told Parler, a popular platform for conservatives, to step up its policing to stay in their app stores."
#> [37] "Can a president be impeached in 12 days? Here’s how the process might work."                                             
#> [38] "In Capital, a G.O.P. Crisis. At the R.N.C. Meeting, a Trump Celebration."                                                
#> [39] "Senator Josh Hawley, who drew condemnation for challenging the election results, defended his decision."                 
#> [40] "Seeing the Confederate flag in the Capitol was a jarring first in U.S. history. Historians weighed in on the moment."    
#> [41] "For those who survived the Nazi death camp, pictures of a man in a “Camp Auschwitz” sweatshirt were painful."            
#> [42] "Opinion"                                                                                                                 
#> [43] "Editors’ Picks"                                                                                                          
#> [44] "Advertisement"